// frontend/src/app/api/generated/model/nursing-chart-view.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NursingChartView { 
    uid?: string;
    admissionUid?: string;
    nurseUid?: string;
    feeding?: string;
    changingPosition?: string;
    bedBathing?: string;
    randomBloodSugar?: string;
    fullBloodSugar?: string;
    drainageOutput?: string;
    fluidIntake?: string;
    urineOutput?: string;
    createdAt?: string;
}